+++ aqa_2/src/data/locators.ts
