+++ aqa_2/src/data/messages.ts
@@ -1,0 +1,12 @@
+
+export enum LOGIN_MESSAGES_HEROKUAPP {
+    LOGIN_SUCCESS = 'You logged into a secure area!',
+    LOGIN_FAILED = 'Your username is invalid!',
+    LOGOUT = 'You logged out of the secure area!'
+}
+
+export const loginMessages_karpovich = {
+    loginSuccess: "Hello, ",
+    registrationSuccess: "Successfully registered! Please, click Back to return on login page",
+};
+
